+++ src/scl/utility/hash.h
@@ -12,6 +12,7 @@
  * @{
  */
 
+#include <scl/utility/hash/concepts.h>
 #include <scl/utility/hash/djb2.h>
 #include <scl/utility/hash/fnv1a.h>
 #include <scl/utility/hash/jenkins_ota.h>